+++ back/GeneralAPI/GeneralAPI/src/main/java/net/weg/userapi/model/dto/request/annotation/AnnotationStudentRequestDTO.java
@@ -17,9 +17,9 @@ public class AnnotationStudentRequestDTO {
     @NotBlank(message = "{not.blank.message}")
     private String toImprove;
     @NotNull(message = "{not.null.message}")
-    private Integer teacher_id;
+    private Long teacher_id;
     @NotNull(message = "{not.null.message}")
-    private Integer council_id;
+    private Long council_id;
     @NotNull(message = "{not.null.message}")
-    private Integer student_id;
+    private Long student_id;
 }
