+++ back/GeneralAPI/GeneralAPI/src/main/java/net/weg/userapi/model/dto/request/feedback/FeedbackUserRequestDTO.java
@@ -19,12 +19,12 @@ public class FeedbackUserRequestDTO {
     @NotBlank(message = "{not.blank.message}")
     private String rank;
     @NotNull(message = "{not.null.message}")
-    private Integer council_id;
+    private Long council_id;
     @NotBlank(message = "{not.blank.message}")
     private String strengths;
     @NotBlank(message = "{not.blank.message}")
     private String toImprove;
     @NotNull(message = "{not.null.message}")
-    private Integer user_id;
+    private Long user_id;
 
 }
