+++ back/GeneralAPI/GeneralAPI/src/main/java/net/weg/userapi/repository/AdminRepository.java
@@ -5,5 +5,5 @@ import net.weg.userapi.model.entity.users.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface AdminRepository extends JpaRepository<Admin, Integer>, JpaSpecificationExecutor<Admin> {
+public interface AdminRepository extends JpaRepository<Admin, Long>, JpaSpecificationExecutor<Admin> {
 }
