+++ back/GeneralAPI/GeneralAPI/src/main/java/net/weg/userapi/repository/AnnotationClassRepository.java
@@ -5,5 +5,5 @@ import net.weg.userapi.model.entity.classes.Class;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface AnnotationClassRepository extends JpaRepository<AnnotationClass, Integer>, JpaSpecificationExecutor<AnnotationClass> {
+public interface AnnotationClassRepository extends JpaRepository<AnnotationClass, Long>, JpaSpecificationExecutor<AnnotationClass> {
 }
