+++ back/GeneralAPI/GeneralAPI/src/main/java/net/weg/userapi/repository/ClassRepository.java
@@ -4,5 +4,5 @@ import net.weg.userapi.model.entity.classes.Class;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ClassRepository extends JpaRepository<Class, Integer>, JpaSpecificationExecutor<Class> {
+public interface ClassRepository extends JpaRepository<Class, Long>, JpaSpecificationExecutor<Class> {
 }
