+++ back/GeneralAPI/GeneralAPI/src/main/java/net/weg/userapi/repository/PreCouncilRepository.java
@@ -5,5 +5,5 @@ import net.weg.userapi.model.entity.preCouncil.PreCouncil;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface PreCouncilRepository extends JpaRepository<PreCouncil, Integer>, JpaSpecificationExecutor<PreCouncil> {
+public interface PreCouncilRepository extends JpaRepository<PreCouncil, Long>, JpaSpecificationExecutor<PreCouncil> {
 }
