+++ back/GeneralAPI/GeneralAPI/src/main/java/net/weg/userapi/repository/PreCouncilSectionRepository.java
@@ -5,5 +5,5 @@ import net.weg.userapi.model.entity.preCouncil.PreCouncilSection;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface PreCouncilSectionRepository extends JpaRepository<PreCouncilSection, Integer>, JpaSpecificationExecutor<PreCouncilSection> {
+public interface PreCouncilSectionRepository extends JpaRepository<PreCouncilSection, Long>, JpaSpecificationExecutor<PreCouncilSection> {
 }
