+++ back/GeneralAPI/GeneralAPI/src/main/java/net/weg/userapi/repository/TeacherRepository.java
@@ -5,5 +5,5 @@ import net.weg.userapi.model.entity.users.Teacher;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface TeacherRepository extends JpaRepository<Teacher, Integer>, JpaSpecificationExecutor<Teacher> {
+public interface TeacherRepository extends JpaRepository<Teacher, Long>, JpaSpecificationExecutor<Teacher> {
 }
